requirements.txt
azure-functions==1.0.0a4
azure-functions-worker==1.0.0a4
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.6.1
six==1.11.0
tensorflow==1.5.0
pillow==5.0.0
numpy==1.14.1
flask
urllib3